moving delta sharing resources to sharing namepsace
