+++ provider/provider.go
@@ -84,8 +84,8 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_schemas":                  catalog.DataSourceSchemas().ToResource(),
 			"databricks_service_principal":        scim.DataSourceServicePrincipal().ToResource(),
 			"databricks_service_principals":       scim.DataSourceServicePrincipals().ToResource(),
-			"databricks_share":                    catalog.DataSourceShare().ToResource(),
-			"databricks_shares":                   catalog.DataSourceShares().ToResource(),
+			"databricks_share":                    sharing.DataSourceShare().ToResource(),
+			"databricks_shares":                   sharing.DataSourceShares().ToResource(),
 			"databricks_spark_version":            clusters.DataSourceSparkVersion().ToResource(),
 			"databricks_sql_warehouse":            sql.DataSourceWarehouse().ToResource(),
 			"databricks_sql_warehouses":           sql.DataSourceWarehouses().ToResource(),
@@ -151,7 +151,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_permission_assignment":       access.ResourcePermissionAssignment().ToResource(),
 			"databricks_permissions":                 permissions.ResourcePermissions().ToResource(),
 			"databricks_pipeline":                    pipelines.ResourcePipeline().ToResource(),
-			"databricks_provider":                    catalog.ResourceProvider().ToResource(),
+			"databricks_provider":                    sharing.ResourceProvider().ToResource(),
 			"databricks_recipient":                   sharing.ResourceRecipient().ToResource(),
 			"databricks_registered_model":            catalog.ResourceRegisteredModel().ToResource(),
 			"databricks_repo":                        repos.ResourceRepo().ToResource(),
@@ -162,7 +162,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_service_principal":           scim.ResourceServicePrincipal().ToResource(),
 			"databricks_service_principal_role":      aws.ResourceServicePrincipalRole().ToResource(),
 			"databricks_service_principal_secret":    tokens.ResourceServicePrincipalSecret().ToResource(),
-			"databricks_share":                       catalog.ResourceShare().ToResource(),
+			"databricks_share":                       sharing.ResourceShare().ToResource(),
 			"databricks_sql_dashboard":               sql.ResourceSqlDashboard().ToResource(),
 			"databricks_sql_endpoint":                sql.ResourceSqlEndpoint().ToResource(),
 			"databricks_sql_global_config":           sql.ResourceSqlGlobalConfig().ToResource(),
