+++ sharing/data_shares.go
@@ -1,4 +1,4 @@
-package catalog
+package sharing
 
 import (
 	"context"
